Fix stored XSS via file upload
Restrict allowed extensions in file upload and send `X-Content-Type-Options: nosniff`. XXS was possible by:
1. Uploading HTML as a part
2. Uploading HTML as a part image
3. Uploading HTML as a profile picture
4. Uploading malicious SVG as an image

Now file uploads are verified on the server. Extension or mime type deduced from the extension (NOT `Content-Type` which can be manipulated) is compared with a (short) list of allowed extensions. `nosniff` is set to prevent content sniffing and finding HTML/JS where it shouldn't be.

Additionally, the `save_image` function was modified to add the file extension, because multiple image types are allowed.