Change how recaptcha is invoked
For some reason my callback `onSubmitCaptcha` would not be called for the second time if the button was clicked, so if the user filled the form wrong and was prompted by the browser, they would be unable to submit it after correcting. Probably when the captcha expired (2 min), it would work again, so now our script handles form submissions and invokes captcha manually.